refactor: adds useSelectedConfigureState manages the configure state of the selected tab
