+++ src/components/Configure/content/fields/RequiredFields.tsx
@@ -1,17 +1,13 @@
 import { Box, Tag } from '@chakra-ui/react';
 
 import { useProject } from '../../../../context/ProjectContext';
-import { useSelectedObjectName } from '../../ObjectManagementNav';
-import { useConfigureState } from '../../state/ConfigurationStateProvider';
-import { getConfigureState } from '../../state/utils';
 import { isIntegrationFieldMapping } from '../../utils';
+import { useSelectedConfigureState } from '../useSelectedConfigureState';
 
 import { FieldHeader } from './FieldHeader';
 
 export function RequiredFields() {
-  const { selectedObjectName } = useSelectedObjectName();
-  const { objectConfigurationsState } = useConfigureState();
-  const configureState = getConfigureState(selectedObjectName || '', objectConfigurationsState);
+  const { configureState, selectedObjectName } = useSelectedConfigureState();
   const { appName } = useProject();
 
   return (
